UPSTREAM: drm/i915: fix the __intel_fbc_update() comments
Don't try to list in comments the cases where we should enable or
disable FBC: it varies a lot with the hardware generations and the
code should be the documentation. Also notice that there's already a
huge gap between the comments and what's in the code.

Signed-off-by: Paulo Zanoni <paulo.r.zanoni@intel.com>

Link: http://patchwork.freedesktop.org/patch/msgid/1446664257-32012-7-git-send-email-paulo.r.zanoni@intel.com
Signed-off-by: Maarten Lankhorst <maarten.lankhorst@linux.intel.com>
(cherry picked from commit 548043abaea4b078270025205833315a2d767e86
from v4.5 )
Signed-off-by: Tarun Vyas <tarun.vyas@intel.com>

Change-Id: Ib1e0e5ae8c5e4966d2c4cea4cdab0c3badc5a0ac