UPSTREAM: ASoC: ac97: add gpio chip
The AC97 specification provides a guide for 16 GPIOs in the codecs. If
the gpiolib is compiled in the kernel, declare a gpio chip.

This was tested with a pxa27x board (mioa701) and a wm9713 codec.

Signed-off-by: Robert Jarzmik <robert.jarzmik@free.fr>
Signed-off-by: Mark Brown <broonie@kernel.org>
(cherry picked from commit 9bf5c3d11f1fbaf43399d189f05fb20ceb46ee5d
from v4.5)
Signed-off-by: Tarun Vyas <tarun.vyas@intel.com>

Change-Id: I1d90b76cdbcea32236f44e7ca0454460900b99c9