UPSTREAM: drm/i915: Simplify checking of GPU reset_counter in display pageflips
If we, when we store the reset_counter for the operation, we ensure that
it is not in a wedged or in the middle of a reset, we can then assert that
if any reset occurs the reset_counter must change. Later we can just
compare the operation's reset epoch against the current counter to see
if we need to abort the operation (to handle the hang).

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Daniel Vetter <daniel.vetter@ffwll.ch>
Reviewed-by: Daniel Vetter <daniel.vetter@ffwll.ch>
Link: http://patchwork.freedesktop.org/patch/msgid/1460565315-7748-5-git-send-email-chris@chris-wilson.co.uk
(cherry picked from commit 7f1847ebf48b2e5753691cc9e2a404c260383933
from drm-intel-next-2016-04-25)
Signed-off-by: Tarun Vyas <tarun.vyas@intel.com>

BUG=chrome-os-partner:52640
TEST=Image builds, boots and is browse-able

Change-Id: Ie654a9b9abc2fe5b3095e49cd22ba5411c6795cb
Reviewed-on: https://chromium-review.googlesource.com/345260
Commit-Ready: Tarun Vyas <tarun.vyas@intel.com>
Tested-by: Russ Sage <russ.sage@intel.com>
Reviewed-by: Stéphane Marchesin <marcheu@chromium.org>