drm: fix up post-merge commit
It appears that the merge of drm-forklift-v4.4.14 into
standard/intel/base was completed without issue, but there
were errors with the merge conflict resolution when merging base
into bxt-rebase.  This patch applies the diff between standard/intel/bxt-rebase
and standard/intel/base for drivers/gpu/drm and include/drm.

After this, compilation of bxt-rebase is successful, and the drm driver
matches what is expected in drm-forklift-v4.4.14 exactly.

Signed-off-by: Eric Ernst <eric.ernst@linux.intel.com>
Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>